Arreglo promedio meme inicio

+++ backend/src/db/memes.js
@@ -19,7 +19,7 @@ async function getAllMemes() {
       m.descripcion,
       m.imagen_url,
       m.fecha_publicacion,
-    COALESCE(AVG(p.puntaje), 0) AS promedio_puntaje, 
+    COALESCE(AVG(p.puntaje)::float, 0) AS promedio_puntaje, 
     COUNT(DISTINCT c.id_comentario) AS cantidad_comentarios
     FROM memes m
     LEFT JOIN puntuaciones_memes p ON m.id_meme = p.meme_id
